+++ packages/xchain-cosmos/src/client.ts
@@ -1,4 +1,4 @@
-import { proto } from '@cosmos-client/core'
+import { cosmosclient, proto } from '@cosmos-client/core'
 import {
   Address,
   Balance,
@@ -13,13 +13,16 @@ import {
   TxsPage,
   XChainClient,
   XChainClientParams,
+  singleFee,
 } from '@xchainjs/xchain-client'
-import { Asset, Chain, assetToString, baseAmount } from '@xchainjs/xchain-util'
+import { Asset, Chain, baseAmount, eqAsset } from '@xchainjs/xchain-util'
+import BigNumber from 'bignumber.js'
 
+import { DECIMAL, DEFAULT_FEE, DEFAULT_GAS_LIMIT } from './const'
 import { CosmosSDKClient } from './cosmos/sdk-client'
 import { TxOfflineParams } from './cosmos/types'
-import { AssetAtom, AssetMuon } from './types'
-import { DECIMAL, getAsset, getDenom, getTxsFromHistory } from './util'
+import { AssetAtom, AssetMuon, ChainIds, ClientUrls, CosmosClientParams } from './types'
+import { getAsset, getDefaultChainIds, getDefaultClientUrls, getDenom, getTxsFromHistory } from './util'
 
 /**
  * Interface for custom Cosmos client
@@ -29,20 +32,13 @@ export interface CosmosClient {
   getSDKClient(): CosmosSDKClient
 }
 
-const MAINNET_SDK = new CosmosSDKClient({
-  server: 'https://api.cosmos.network',
-  chainId: 'cosmoshub-4',
-})
-const TESTNET_SDK = new CosmosSDKClient({
-  server: 'https://rest.sentry-02.theta-testnet.polypore.xyz',
-  chainId: 'theta-testnet-001',
-})
-
 /**
  * Custom Cosmos client
  */
 class Client extends BaseXChainClient implements CosmosClient, XChainClient {
-  private sdkClients: Map<Network, CosmosSDKClient> = new Map<Network, CosmosSDKClient>()
+  private sdkClient: CosmosSDKClient
+  private clientUrls: ClientUrls
+  private chainIds: ChainIds
 
   /**
    * Constructor
@@ -57,15 +53,40 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
   constructor({
     network = Network.Testnet,
     phrase,
+    clientUrls,
+    chainIds,
     rootDerivationPaths = {
       [Network.Mainnet]: `44'/118'/0'/0/`,
       [Network.Testnet]: `44'/118'/0'/0/`,
       [Network.Stagenet]: `44'/118'/0'/0/`,
     },
-  }: XChainClientParams) {
+  }: XChainClientParams & CosmosClientParams) {
     super(Chain.Cosmos, { network, rootDerivationPaths, phrase })
-    this.sdkClients.set(Network.Testnet, TESTNET_SDK)
-    this.sdkClients.set(Network.Mainnet, MAINNET_SDK)
+
+    this.clientUrls = clientUrls || getDefaultClientUrls()
+    this.chainIds = chainIds || getDefaultChainIds()
+
+    this.sdkClient = new CosmosSDKClient({
+      server: this.clientUrls[network],
+      chainId: this.chainIds[network],
+    })
+  }
+
+  /**
+   * Updates current network.
+   *
+   * @param {Network} network
+   * @returns {void}
+   */
+  setNetwork(network: Network): void {
+    // dirty check to avoid using and re-creation of same data
+    if (network === this.network) return
+
+    super.setNetwork(network)
+    this.sdkClient = new CosmosSDKClient({
+      server: this.clientUrls[network],
+      chainId: this.chainIds[network],
+    })
   }
 
   /**
@@ -119,7 +140,7 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
   }
 
   getSDKClient(): CosmosSDKClient {
-    return this.sdkClients.get(this.network) || TESTNET_SDK
+    return this.sdkClient
   }
 
   /**
@@ -168,19 +189,16 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
    * @returns {Balance[]} The balance of the address.
    */
   async getBalance(address: Address, assets?: Asset[]): Promise<Balance[]> {
-    const balances = await this.getSDKClient().getBalance(address)
-    const mainAsset = this.getMainAsset()
+    const coins = await this.getSDKClient().getBalance(address)
+
+    const balances = coins
+      .reduce((acc: Balance[], { denom, amount }) => {
+        const asset = getAsset(denom)
+        return asset ? [...acc, { asset, amount: baseAmount(amount || '0', DECIMAL) }] : acc
+      }, [])
+      .filter(({ asset: balanceAsset }) => !assets || assets.filter((asset) => eqAsset(balanceAsset, asset)).length)
 
     return balances
-      .map((balance) => {
-        return {
-          asset: (balance.denom && getAsset(balance.denom)) || mainAsset,
-          amount: baseAmount(balance.amount, DECIMAL),
-        }
-      })
-      .filter(
-        (balance) => !assets || assets.filter((asset) => assetToString(balance.asset) === assetToString(asset)).length,
-      )
   }
 
   /**
@@ -238,10 +256,23 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
    * @param {TxParams} params The transfer options.
    * @returns {TxHash} The transaction hash.
    */
-  async transfer({ walletIndex, asset, amount, recipient, memo }: TxParams): Promise<TxHash> {
+  async transfer({
+    walletIndex,
+    asset,
+    amount,
+    recipient,
+    memo,
+    gasLimit = new BigNumber(DEFAULT_GAS_LIMIT),
+  }: TxParams & { gasLimit?: BigNumber }): Promise<TxHash> {
     const fromAddressIndex = walletIndex || 0
 
     const mainAsset = this.getMainAsset()
+
+    const fee = new proto.cosmos.tx.v1beta1.Fee({
+      amount: [],
+      gas_limit: cosmosclient.Long.fromString(gasLimit.toFixed(0)),
+    })
+
     return this.getSDKClient().transfer({
       privkey: this.getPrivateKey(fromAddressIndex),
       from: this.getAddress(fromAddressIndex),
@@ -249,6 +280,7 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
       amount: amount.amount().toString(),
       asset: getDenom(asset || mainAsset),
       memo,
+      fee,
     })
   }
 
@@ -266,10 +298,15 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
     memo,
     from_account_number,
     from_sequence,
+    gasLimit = new BigNumber(DEFAULT_GAS_LIMIT),
   }: TxOfflineParams): Promise<string> {
     const fromAddressIndex = walletIndex || 0
 
-    const mainAsset = this.getMainAsset()
+    const fee = new proto.cosmos.tx.v1beta1.Fee({
+      amount: [],
+      gas_limit: cosmosclient.Long.fromString(gasLimit.toFixed(0)),
+    })
+
     return await this.getSDKClient().transferSignedOffline({
       privkey: this.getPrivateKey(fromAddressIndex),
       from: this.getAddress(fromAddressIndex),
@@ -277,24 +314,19 @@ class Client extends BaseXChainClient implements CosmosClient, XChainClient {
       from_sequence,
       to: recipient,
       amount: amount.amount().toString(),
-      asset: getDenom(asset || mainAsset),
+      asset: getDenom(asset || this.getMainAsset()),
       memo,
+      fee,
     })
   }
 
   /**
-   * Get the current fee.
+   * Returns (default) fees.
    *
-   * @returns {Fees} The current fee.
+   * @returns {Fees} Current fees
    */
   async getFees(): Promise<Fees> {
-    // there is no fixed fee, we set fee amount when creating a transaction.
-    return {
-      type: FeeType.FlatFee,
-      fast: baseAmount(750, DECIMAL),
-      fastest: baseAmount(2500, DECIMAL),
-      average: baseAmount(0, DECIMAL),
-    }
+    return singleFee(FeeType.FlatFee, DEFAULT_FEE)
   }
 }