+++ packages/xchain-cosmos/src/index.ts
@@ -1,4 +1,5 @@
+export * from './types'
+export * from './const'
+export * from './util'
 export * from './cosmos'
 export * from './client'
-export * from './types'
-export * from './util'